+++ src/lib/djerba/util/oncokb/constants.py
@@ -61,6 +61,6 @@ UPDATE_CACHE = 'update cache'
 
 ### miscellaneous ###
 
-ALL_CURATED_GENES = '20231116-allCuratedGenes.tsv'
+ALL_CURATED_GENES = '20240315-allCuratedGenes.tsv'
 ONCOGENIC_UC = 'ONCOGENIC'
 DEFAULT_CACHE_PATH = '/.mounts/labs/CGI/gsi/tools/djerba/oncokb_cache/scratch'