+++ CHANGELOG.md
@@ -11,9 +11,15 @@
 - GCGI-1190: Removed obsolete data values/files
 - GCGI-1012: Moved R testing to formal tests
 - GCGI-1304: Merge updates to genomic landscape plugin and resolve conflicts
+- GCGI-1319: Update genomic landscape plugin to read purity from PURPLE output file
 
 - new PURPLE CNV calling plugin
 - removed sequenza support
+- new `djerba.plugins.wgts.common` package for code shared between multiple WGTS/WGS plugins
+
+### Other
+
+- GCGI-1313: Refactor reading input_params.json
 
 ## v1.5.5: 2024-03-05
 - Further fixes to pWGS Cardea helper to allow multiple projects for one donor